requirement.txt
blis==0.4.1
boto3==1.10.34
botocore==1.13.34
catalogue==0.0.8
certifi==2019.11.28
chardet==3.0.4
cymem==2.0.3
Django==2.2.8
djangorestframework==3.10.3
docutils==0.15.2
idna==2.8
importlib-metadata==1.2.0
jmespath==0.9.4
more-itertools==8.0.2
murmurhash==1.0.2
numpy==1.17.4
pandas==0.24.2
pkg-resources==0.0.0
plac==1.1.3
preshed==3.0.2
pydub==0.23.1
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
s3transfer==0.2.1
six==1.13.0
spacy==2.2.3
SpeechRecognition==3.8.1
sqlparse==0.3.0
srsly==0.2.0
thinc==7.3.1
tqdm==4.40.1
urllib3==1.25.7
vectors==1.0.0
wasabi==0.4.2
zipp==0.6.0
